+++ pyqir-generator/src/qir/instructions.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-use super::{array1d, basic_values, calls, result};
+use super::result;
 use crate::interop::{If, Instruction};
 use inkwell::values::{BasicValueEnum, FunctionValue, PointerValue};
 use qirlib::codegen::CodeGenerator;
@@ -11,49 +11,12 @@ use std::collections::HashMap;
 ///
 /// Panics if the qubit name doesn't exist
 fn get_qubit<'ctx>(
-    qubits: &HashMap<String, BasicValueEnum<'ctx>>,
     name: &str,
+    qubits: &HashMap<String, BasicValueEnum<'ctx>>,
 ) -> BasicValueEnum<'ctx> {
     *qubits
         .get(name)
         .unwrap_or_else(|| panic!("Qubit {} not found.", name))
-}
-
-fn get_register_result_pointer<'a>(
-    generator: &CodeGenerator<'a>,
-    registers: &HashMap<String, (BasicValueEnum<'a>, Option<u64>)>,
-    name: &str,
-) -> PointerValue<'a> {
-    let (register, index) = registers
-        .get(name)
-        .unwrap_or_else(|| panic!("Register {} not found.", name));
-
-    array1d::get_bitcast_result_pointer_array_element(generator, index.unwrap(), register, "")
-        .into_pointer_value()
-}
-
-fn get_register_result<'a>(
-    generator: &CodeGenerator<'a>,
-    registers: &HashMap<String, (BasicValueEnum<'a>, Option<u64>)>,
-    name: &str,
-) -> PointerValue<'a> {
-    generator
-        .builder
-        .build_load(get_register_result_pointer(generator, registers, name), "")
-        .into_pointer_value()
-}
-
-fn set_register_result<'a>(
-    generator: &CodeGenerator<'a>,
-    registers: &HashMap<String, (BasicValueEnum<'a>, Option<u64>)>,
-    name: &str,
-    result: PointerValue<'a>,
-) {
-    let pointer = get_register_result_pointer(generator, registers, name);
-    let old_result = generator.builder.build_load(pointer, "");
-    result::update_reference_count(generator, old_result, -1);
-    result::update_reference_count(generator, result, 1);
-    generator.builder.build_store(pointer, result);
 }
 
 fn measure<'ctx>(
@@ -61,22 +24,18 @@ fn measure<'ctx>(
     qubit: &str,
     target: &str,
     qubits: &HashMap<String, BasicValueEnum<'ctx>>,
-    registers: &HashMap<String, (BasicValueEnum<'ctx>, Option<u64>)>,
+    registers: &mut HashMap<String, Option<PointerValue<'ctx>>>,
 ) {
-    let m = generator
-        .intrinsics
-        .m
-        .expect("m must be defined in the template");
+    let find_qubit = |name| get_qubit(name, qubits);
 
-    let result = calls::emit_call_with_return(
-        &generator.builder,
-        m,
-        &[get_qubit(qubits, qubit).into()],
-        "measurement",
-    )
-    .into_pointer_value();
+    // measure the qubit and save the result to a temporary value
+    let new_value = generator.emit_call_with_return(
+        generator.qis_m_body(),
+        &[find_qubit(qubit).into()],
+        target,
+    );
 
-    set_register_result(generator, registers, target, result);
+    registers.insert(target.to_owned(), Some(new_value.into_pointer_value()));
 }
 
 fn controlled<'ctx>(
@@ -85,18 +44,7 @@ fn controlled<'ctx>(
     control: BasicValueEnum<'ctx>,
     qubit: BasicValueEnum<'ctx>,
 ) {
-    calls::emit_void_call(
-        &generator.builder,
-        intrinsic,
-        &[control.into(), qubit.into()],
-    );
-
-    let minus_one = basic_values::i64_to_i32(generator, -1);
-    generator.builder.build_call(
-        generator.runtime_library.array_update_reference_count,
-        &[control.into(), minus_one],
-        "",
-    );
+    generator.emit_void_call(intrinsic, &[control.into(), qubit.into()]);
 }
 
 #[allow(clippy::too_many_lines)]
@@ -104,138 +52,99 @@ pub(crate) fn emit<'ctx>(
     generator: &CodeGenerator<'ctx>,
     inst: &Instruction,
     qubits: &HashMap<String, BasicValueEnum<'ctx>>,
-    registers: &HashMap<String, (BasicValueEnum<'ctx>, Option<u64>)>,
+    registers: &mut HashMap<String, Option<PointerValue<'ctx>>>,
 ) {
-    let intrinsics = &generator.intrinsics;
-    let find_qubit = |name| get_qubit(qubits, name);
-    let ctl = |value| array1d::create_ctl_wrapper(generator, value);
-
+    let find_qubit = |name| get_qubit(name, qubits);
     match inst {
         Instruction::Cx(inst) => {
-            let control = ctl(&find_qubit(&inst.control));
+            let control = find_qubit(&inst.control);
             let qubit = find_qubit(&inst.target);
-            controlled(
-                generator,
-                intrinsics
-                    .x_ctl
-                    .expect("x_ctl must be defined in the template"),
-                control,
-                qubit,
-            );
+            controlled(generator, generator.qis_cnot_body(), control, qubit);
         }
         Instruction::Cz(inst) => {
-            let control = ctl(&find_qubit(&inst.control));
+            let control = find_qubit(&inst.control);
             let qubit = find_qubit(&inst.target);
-            controlled(
-                generator,
-                intrinsics
-                    .z_ctl
-                    .expect("z_ctl must be defined in the template"),
-                control,
-                qubit,
+            controlled(generator, generator.qis_cz_body(), control, qubit);
+        }
+        Instruction::H(inst) => {
+            generator.emit_void_call(generator.qis_h_body(), &[find_qubit(&inst.qubit).into()]);
+        }
+        Instruction::M(inst) => {
+            measure(generator, &inst.qubit, &inst.target, qubits, registers);
+        }
+        Instruction::Reset(inst) => {
+            generator.emit_void_call(
+                generator.qis_reset_body(),
+                &[find_qubit(&inst.qubit).into()],
             );
         }
-        Instruction::H(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.h.expect("h must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::M(inst) => measure(generator, &inst.qubit, &inst.target, qubits, registers),
-        Instruction::Reset(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics
-                .reset
-                .expect("reset must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::Rx(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.r_x.expect("r_x must be defined in the template"),
-            &[
-                basic_values::f64_to_f64(generator, inst.theta),
-                find_qubit(&inst.qubit).into(),
-            ],
-        ),
-        Instruction::Ry(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.r_y.expect("r_y must be defined in the template"),
-            &[
-                basic_values::f64_to_f64(generator, inst.theta),
-                find_qubit(&inst.qubit).into(),
-            ],
-        ),
-        Instruction::Rz(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.r_z.expect("r_z must be defined in the template"),
-            &[
-                basic_values::f64_to_f64(generator, inst.theta),
-                find_qubit(&inst.qubit).into(),
-            ],
-        ),
-        Instruction::S(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.s.expect("s must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::SAdj(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics
-                .s_adj
-                .expect("s_adj must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::T(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.t.expect("t must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::TAdj(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics
-                .t_adj
-                .expect("t_adj must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::X(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.x.expect("x must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::Y(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.y.expect("y must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::Z(inst) => calls::emit_void_call(
-            &generator.builder,
-            intrinsics.z.expect("z must be defined in the template"),
-            &[find_qubit(&inst.qubit).into()],
-        ),
-        Instruction::DumpMachine => calls::emit_void_call(
-            &generator.builder,
-            intrinsics
-                .dumpmachine
-                .expect("dumpmachine must be defined before use"),
-            &[basic_values::i8_null_ptr(generator)],
-        ),
+        Instruction::Rx(inst) => {
+            generator.emit_void_call(
+                generator.qis_rx_body(),
+                &[
+                    generator.f64_to_f64(inst.theta),
+                    find_qubit(&inst.qubit).into(),
+                ],
+            );
+        }
+        Instruction::Ry(inst) => {
+            generator.emit_void_call(
+                generator.qis_ry_body(),
+                &[
+                    generator.f64_to_f64(inst.theta),
+                    find_qubit(&inst.qubit).into(),
+                ],
+            );
+        }
+        Instruction::Rz(inst) => {
+            generator.emit_void_call(
+                generator.qis_rz_body(),
+                &[
+                    generator.f64_to_f64(inst.theta),
+                    find_qubit(&inst.qubit).into(),
+                ],
+            );
+        }
+        Instruction::S(inst) => {
+            generator.emit_void_call(generator.qis_s_body(), &[find_qubit(&inst.qubit).into()]);
+        }
+        Instruction::SAdj(inst) => {
+            generator.emit_void_call(generator.qis_s_adj(), &[find_qubit(&inst.qubit).into()]);
+        }
+        Instruction::T(inst) => {
+            generator.emit_void_call(generator.qis_t_body(), &[find_qubit(&inst.qubit).into()]);
+        }
+        Instruction::TAdj(inst) => {
+            generator.emit_void_call(generator.qis_t_adj(), &[find_qubit(&inst.qubit).into()]);
+        }
+        Instruction::X(inst) => {
+            generator.emit_void_call(generator.qis_x_body(), &[find_qubit(&inst.qubit).into()]);
+        }
+        Instruction::Y(inst) => {
+            generator.emit_void_call(generator.qis_y_body(), &[find_qubit(&inst.qubit).into()]);
+        }
+        Instruction::Z(inst) => {
+            generator.emit_void_call(generator.qis_z_body(), &[find_qubit(&inst.qubit).into()]);
+        }
         Instruction::If(if_inst) => emit_if(generator, registers, qubits, if_inst),
     }
 }
 
-fn emit_if<'a>(
-    generator: &CodeGenerator<'a>,
-    registers: &HashMap<String, (BasicValueEnum<'a>, Option<u64>)>,
-    qubits: &HashMap<String, BasicValueEnum<'a>>,
+fn emit_if<'ctx>(
+    generator: &CodeGenerator<'ctx>,
+    registers: &mut HashMap<String, Option<PointerValue<'ctx>>>,
+    qubits: &HashMap<String, BasicValueEnum<'ctx>>,
     if_inst: &If,
 ) {
     // TODO: Refactor this with qir::get_entry_function.
     let entry_name = "QuantumApplication__Run__body";
     let entry = generator.module.get_function(entry_name).unwrap();
 
-    // The reference count doesn't need to be updated because the result is only used for this
-    // condition and won't outlive the array.
-    let result = get_register_result(generator, registers, &if_inst.condition);
-    let condition = result::equals(generator, result, result::get_one(generator));
+    let result = registers
+        .get(&if_inst.condition)
+        .and_then(|v| *v)
+        .unwrap_or_else(|| panic!("Result {} not found.", &if_inst.condition));
+    let condition = result::equal(generator, result, result::get_one(generator));
 
     let then_block = generator.context.append_basic_block(entry, "then");
     let else_block = generator.context.append_basic_block(entry, "else");
@@ -244,7 +153,7 @@ fn emit_if<'a>(
         .build_conditional_branch(condition, then_block, else_block);
 
     let continue_block = generator.context.append_basic_block(entry, "continue");
-    let emit_block = |block, insts| {
+    let mut emit_block = |block, insts| {
         generator.builder.position_at_end(block);
         for inst in insts {
             emit(generator, inst, qubits, registers);