+++ pyqir-generator/src/qir/result.rs
@@ -1,43 +1,20 @@
-use super::{basic_values, calls};
 use inkwell::values::{BasicMetadataValueEnum, IntValue, PointerValue};
 use qirlib::codegen::CodeGenerator;
 
 pub(crate) fn get_one<'a>(generator: &CodeGenerator<'a>) -> PointerValue<'a> {
-    calls::emit_call_with_return(
-        &generator.builder,
-        generator.runtime_library.result_get_one,
-        &[],
-        "one",
-    )
-    .into_pointer_value()
+    generator
+        .emit_call_with_return(generator.rt_result_get_one(), &[], "one")
+        .into_pointer_value()
 }
 
-pub(crate) fn equals<'a>(
+pub(crate) fn equal<'a>(
     generator: &CodeGenerator<'a>,
-    x: PointerValue<'a>,
-    y: PointerValue<'a>,
+    result1: PointerValue<'a>,
+    result2: PointerValue<'a>,
 ) -> IntValue<'a> {
-    let x = BasicMetadataValueEnum::PointerValue(x);
-    let y = BasicMetadataValueEnum::PointerValue(y);
-
-    calls::emit_call_with_return(
-        &generator.builder,
-        generator.runtime_library.result_equal,
-        &[x, y],
-        "",
-    )
-    .into_int_value()
-}
-
-pub(crate) fn update_reference_count<'a>(
-    generator: &CodeGenerator,
-    result: impl Into<BasicMetadataValueEnum<'a>>,
-    delta: i32,
-) {
-    let delta = basic_values::i64_to_i32(generator, delta.into());
-    generator.builder.build_call(
-        generator.runtime_library.result_update_reference_count,
-        &[result.into(), delta],
-        "",
-    );
+    let result1 = BasicMetadataValueEnum::PointerValue(result1);
+    let result2 = BasicMetadataValueEnum::PointerValue(result2);
+    generator
+        .emit_call_with_return(generator.rt_result_equal(), &[result1, result2], "equal")
+        .into_int_value()
 }