+++ pyqir-jit/src/interop.rs
@@ -112,7 +112,6 @@ pub enum Instruction {
     X(Single),
     Y(Single),
     Z(Single),
-    DumpMachine,
 }
 
 #[derive(Clone, Default)]