+++ qirlib/src/lib.rs
@@ -4,9 +4,5 @@
 #![warn(clippy::all, clippy::pedantic)]
 
 pub mod codegen;
-pub mod constants;
-pub mod intrinsics;
 pub mod module;
 pub mod passes;
-pub mod runtime_library;
-pub mod types;